Revise dto content population and dynamic request population
