+++ db/queries/instantiations/data_requests/post.py
@@ -3,7 +3,7 @@ from typing import override, final
 from db.models.implementations import LinkLocationDataRequest
 from db.models.implementations.core.data_request.core import DataRequest
 from db.queries.builder.core import QueryBuilderBase
-from middleware.schema_and_dto.dtos.data_requests.post import DataRequestsPostDTO
+from endpoints.instantiations.data_requests_.post.dto import DataRequestsPostDTO
 
 
 @final
