+++ endpoints/psycopg_resource.py
@@ -12,9 +12,8 @@ from db.helpers_.psycopg import initialize_psycopg_connection
 from middleware.schema_and_dto.dynamic.dto_request_content_population import (
     populate_dto_with_request_content,
 )
-from middleware.schema_and_dto.dynamic.schema.request_content_population import (
-    populate_schema_with_request_content,
-)
+from middleware.schema_and_dto.dynamic.schema.request_content_population_.core import \
+    populate_schema_with_request_content
 from middleware.schema_and_dto.non_dto_dataclasses import (
     SchemaPopulateParameters,
     DTOPopulateParameters,
