+++ middleware/primary_resource_logic/data_requests_/post.py
@@ -2,7 +2,7 @@ from flask import Response
 
 from db.client.core import DatabaseClient
 from middleware.common_response_formatting import created_id_response
-from middleware.schema_and_dto.dtos.data_requests.post import DataRequestsPostDTO
+from endpoints.instantiations.data_requests_.post.dto import DataRequestsPostDTO
 from middleware.security.access_info.primary import AccessInfoPrimary
