+++ middleware/schema_and_dto/dynamic/pydantic_to_marshmallow/mapping.py
@@ -1,4 +1,5 @@
 import datetime
+from typing import get_origin
 
 from marshmallow import fields
 
@@ -11,3 +12,12 @@ TYPE_MAPPING = {
     datetime.datetime: fields.DateTime,
     dict: fields.Dict,
 }
+
+
+def _is_mapped_type(inner_type: type) -> bool:
+    available_types = list(TYPE_MAPPING.keys())
+    if get_origin(inner_type) in available_types:
+        return True
+    if inner_type in available_types:
+        return True
+    return False
