+++ middleware/schema_and_dto/dynamic/schema/doc_construction/types_.py
@@ -1,0 +1,7 @@
+from utilities.enums import SourceMappingEnum
+
+PARSER_FIELDS = [
+    SourceMappingEnum.QUERY_ARGS,
+    SourceMappingEnum.PATH,
+    SourceMappingEnum.FILE,
+]
