+++ middleware/schema_and_dto/dynamic/schema/request_content_population_/exceptions.py
@@ -1,0 +1,2 @@
+class InvalidSourceMappingError(Exception):
+    pass
