+++ middleware/schema_and_dto/dynamic/schema/request_content_population_/models/nested_dto.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+
+class NestedDTOInfo(BaseModel):
+    key: str
+    class_: type
