+++ middleware/schema_and_dto/schemas/data_requests/get/base.py
@@ -1,6 +1,6 @@
 from marshmallow import fields
 
-from middleware.schema_and_dto.schemas.data_requests.base import (
+from endpoints.instantiations.data_requests_._shared.schemas.base import (
     DataRequestsSchema,
 )
 from middleware.schema_and_dto.schemas.locations.info.response import (
