+++ middleware/schema_and_dto/schemas/data_requests/put.py
@@ -1,7 +1,7 @@
 from marshmallow import Schema, fields
 
 from middleware.schema_and_dto.dtos.data_requests.put import DataRequestsPutDTO
-from middleware.schema_and_dto.schemas.data_requests.base import (
+from endpoints.instantiations.data_requests_._shared.schemas.base import (
     DataRequestsSchema,
 )
 from middleware.schema_and_dto.util import get_json_metadata
