+++ middleware/schema_and_dto/schemas/data_sources/get/base.py
@@ -1,7 +1,7 @@
 from marshmallow import fields
 
 from middleware.schema_and_dto.schemas.agencies.get.base import AgenciesGetSchema
-from middleware.schema_and_dto.schemas.data_requests.base import (
+from endpoints.instantiations.data_requests_._shared.schemas.base import (
     DataRequestsSchema,
 )
 from middleware.schema_and_dto.schemas.data_sources.expanded import (
